ref(test): move testIssue575 data into fixtures
stop relying on a local ebin dir. Standardize assertions with other decompiler tests.
